[WIP] Update unit tests
